Shutdown should now work if it is the last vmf to compile

+++ CompilePalX/Compilers/ShutdownProcess.cs
@@ -13,12 +13,9 @@ namespace CompilePalX.Compilers
         {
             CompileErrors = new List<Error>();
 
-
-            /**** TODO: NEEDS UNCOMMENTING AND TWEAKING WHEN MULTIPLE MAP PRESETS CAN BE SELECTED FOR COMPILING AT ONCE ****/
-
             // don't run unless it's the last map of the queue
-            /*if (CompilingManager.MapFiles[ConfigurationManager.CurrentPresetMap].File == context.MapFile)
-            {*/
+            if (CompilingManager.MapFiles.Where(x => x.Value.Compile)?.Select(x => x.Value.File).LastOrDefault() == context.MapFile)
+            {
                 CompilePalLogger.LogLine("\nCompilePal - Shutdown");
                 CompilePalLogger.LogLine("The system will shutdown soon.");
                 CompilePalLogger.LogLine("You can cancel this shutdown by using the command \"shutdown -a\"");
@@ -29,7 +26,7 @@ namespace CompilePalX.Compilers
 
                 Process = new Process { StartInfo = startInfo };
                 Process.Start();
-            //}
+            }
         }
     }
 }
